refactor: enhance ProjectDashboardPage with error handling for tasks and issues; implement ProjectsGridTable for project display
