+++ features/projects/components/index.ts
@@ -1,3 +1,4 @@
+export { ProjectsGridTable } from './projects-grid-table';
 export { ProjectEditForm } from './project-edit-form';
 export { ProjectOverview } from './project-overview';
 export { ProjectCharts } from './project-charts';
